move quote tests for new file
